Adds user agent to all HTTP requests
